Safari-609.1.11.1/LayoutTests/media/modern-media-controls/media-controller/media-controller-click-on-video-background-should-pause-fullscreen-expected.txt

Testing that clicking on a fullscreen video's background pauses it.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaControls.querySelector('button.fullscreen') instanceof HTMLButtonElement became true

Entering fullscreen
PASS media.webkitDisplayingFullscreen is true
PASS media.paused is false

Clicking on video background
PASS media.paused is true

PASS successfullyParsed is true

TEST COMPLETE
